test: split exact-tab delivery tests onto the same seam as their source
pty-delivery.test.ts had grown to cover both delivery paths at once. It now
follows the source split: this file owns the canonical path (find the task's
engine session, or create it), exact-tab-delivery.test.ts owns delivery into
one addressed tab — live, absent, a plain shell, or thawed after a host
restart. Also corrects the --respawn help and docs: only a tab Rove holds no
snapshot for replays its frozen launch command; a snapshot-known tab without a
pinned id comes back on a fresh conversation.
