+++ packages/kobe/test/cli/pty-delivery.test.ts
@@ -1,12 +1,12 @@
 /**
- * `pty-delivery.ts` — the bracketed paste that `kobe api` delivery routes
- * through. The engine-key resolver's own tests live in
- * `pty-engine-key.test.ts`.
+ * `pty-delivery.ts` — the CANONICAL delivery path: find the task's engine
+ * session or create it, and the bracketed paste `kobe api` routes through.
+ * The engine-key resolver's own tests live in `pty-engine-key.test.ts`;
+ * delivery into ONE addressed tab is `exact-tab-delivery.test.ts`.
  */
 
 import type { PtySessionInfo } from "@sma1lboy/kobe-daemon/daemon/pty-host"
 import { describe, expect, it } from "vitest"
-import { deliverToExactTab } from "../../src/cli/api/exact-tab-delivery.ts"
 import { deliverHostedPrompt, deliverToKey } from "../../src/cli/api/pty-delivery.ts"
 import { ApiError } from "../../src/cli/api/types.ts"
 
@@ -400,91 +400,6 @@ describe("deliverHostedPrompt", () => {
     )
     expect(err).toBeInstanceOf(ApiError)
     expect((err as ApiError).code).toBe("ENGINE_NOT_RUNNING")
-  })
-})
-
-describe("deliverToExactTab", () => {
-  function rpcWith(sessions: PtySessionInfo[]) {
-    const calls: string[] = []
-    const engine = echoingPeek()
-    const rpc = {
-      request: async <T>(name: string, payload?: unknown): Promise<T> => {
-        calls.push(name)
-        if (name === "pty.list") return { sessions } as T
-        if (name === "pty.peek") return engine.peek() as T
-        if (name === "pty.write") engine.onWrite((payload as { data?: string }).data ?? "")
-        return {} as T
-      },
-    }
-    return { rpc, calls }
-  }
-
-  it("a DIFFERENT vendor's engine in the addressed tab still receives (cross-vendor send)", async () => {
-    const { rpc, calls } = rpcWith([session("t1::tab-2", ["codex"])])
-    const result = await deliverToExactTab(rpc, "t1", "tab-2", "/wt/t1", "go", {
-      engineBin: "claude", // task vendor is claude; tab runs codex
-      snapshot: psWith("codex"),
-    })
-    expect(result).toMatchObject({ session: "t1::tab-2", delivered: true })
-    expect(calls).toContain("pty.write")
-  })
-
-  it("refuses a tab that is a plain shell (ENGINE_NOT_RUNNING, not a paste)", async () => {
-    const { rpc, calls } = rpcWith([session("t1::tab-2", ["/bin/zsh"])])
-    const err = await deliverToExactTab(rpc, "t1", "tab-2", "/wt/t1", "go", {
-      snapshot: psWith("grep something"),
-    }).then(
-      () => null,
-      (e) => e,
-    )
-    expect((err as ApiError).code).toBe("ENGINE_NOT_RUNNING")
-    expect(calls).not.toContain("pty.write")
-  })
-
-  it("refuses a dead/absent tab (TAB_NOT_FOUND)", async () => {
-    const { rpc } = rpcWith([session("t1::tab-2", ["claude"], false)])
-    const err = await deliverToExactTab(rpc, "t1", "tab-2", "/wt/t1", "go").then(
-      () => null,
-      (e) => e,
-    )
-    expect((err as ApiError).code).toBe("TAB_NOT_FOUND")
-  })
-
-  // A pty-host restart leaves every tab thawed-but-dead. TAB_NOT_FOUND for
-  // one of those sent the caller to `pty-list`, which LISTS it — so the
-  // refusal has to be its own code, and reviving has to be reachable.
-  it("distinguishes a freeze-restored tab from an absent one (TAB_RESTORED, not TAB_NOT_FOUND)", async () => {
-    const { rpc, calls } = rpcWith([{ ...session("t1::tab-2", ["claude"], false), restored: true }])
-    const err = await deliverToExactTab(rpc, "t1", "tab-2", "/wt/t1", "go").then(
-      () => null,
-      (e) => e,
-    )
-    expect((err as ApiError).code).toBe("TAB_RESTORED")
-    // Refusal, not a silent revival: nothing was opened and nothing written.
-    expect(calls).not.toContain("pty.open")
-    expect(calls).not.toContain("pty.write")
-  })
-
-  it("respawns a freeze-restored tab and delivers into it when the caller opts in", async () => {
-    const restored = { ...session("t1::tab-2", ["claude"], false), restored: true }
-    const sessions: PtySessionInfo[] = [restored]
-    const { rpc, calls } = rpcWith(sessions)
-    const inner = rpc.request
-    rpc.request = async <T>(name: string, payload?: unknown): Promise<T> => {
-      const out = await inner<T>(name, payload)
-      // The host's respawn-in-place: the same key comes back alive.
-      if (name === "pty.open") sessions[0] = session("t1::tab-2", ["claude"])
-      return name === "pty.open" ? ({ alive: true, respawned: true } as T) : out
-    }
-    const result = await deliverToExactTab(rpc, "t1", "tab-2", "/wt/t1", "go", {
-      engineBin: "claude",
-      snapshot: psWith("claude"),
-      respawn: () => ({ key: "t1::tab-2", command: ["zsh", "-lc", "claude --resume abc"] }),
-    })
-    expect(calls).toContain("pty.open")
-    // `respawned`, never `started`: the SAME tab came back, scrollback and
-    // conversation intact — a new session would be a different claim.
-    expect(result).toMatchObject({ session: "t1::tab-2", delivered: true, respawned: true, started: false })
   })
 })
